refactor(api): rename table module to delta_table
