+++ src/delta_engine/schema.py
@@ -5,7 +5,7 @@ Import table declarations, columns, data types, properties, and key helpers
 from here when defining desired Delta table schemas.
 """
 
-from delta_engine.api.table import DeltaTable, ForeignKey, Self
+from delta_engine.api.delta_table import DeltaTable, ForeignKey, Self
 from delta_engine.application.properties import Property
 from delta_engine.domain.model import (
     Array,
